backend/requirements.txt
docx==0.2.4
fastapi==0.135.3
jose==1.0.0
langgraph==1.1.6
paddleocr==3.4.0
passlib==1.7.4
pdf2image==1.17.0
Pillow==9.0.1
Pillow==12.2.0
pydantic==2.12.5
pydantic_settings==2.13.1
python-dotenv==1.2.2
Requests==2.33.1
slowapi==0.1.9
SQLAlchemy==1.4.31
starlette==1.0.0
